Modifications on: Example.m. and added a new function that adjusts salinity to the mooring data
I have changed the code Example by adding another line to include that new function of salinity, and modified the one that was adjusting oxygen to work with an input of a table instead of the previous way with the file path. In this way, you can execute either the salinity adjustment function or the oxygen function independently, and they will append the corresponding calibrated values.

The caveat is whether you would use the newly adjusted salinity to calculate the oxygen umol kg

I also adjusted the float_in_mooring_footprint, and now it is working

At the moment, for the selection of the target depth of the float_in_mooring_footprint function, you can only select the dissolved oxygen adjusted